requirements.txt
aws-cdk.core==1.67.0
aws-cdk.aws-ec2==1.67.0
aws-cdk.aws_ecs==1.67.0
aws-cdk.aws_ecs_patterns==1.67.0
aws-cdk.aws_ecr==1.67.0
aws-cdk.aws_rds==1.67.0
aws-cdk.aws_s3==1.67.0
aws-cdk.aws_ssm==1.67.0
aws-cdk.aws_route53==1.67.0
aws_cdk.aws_certificatemanager==1.67.0
aws_cdk.aws_elasticache==1.67.0
aws_cdk.aws_logs==1.67.0
aws_cdk.aws_secretsmanager==1.67.0
aws_cdk.aws_elasticloadbalancingv2==1.67.0
aws_cdk.aws_codepipeline_actions==1.67.0
django-environ==0.4.5


